+++ src/client_test.go
@@ -2,6 +2,8 @@ package main
 
 import (
 	"fmt"
+	"net/http"
+	"net/http/httptest"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -64,5 +66,53 @@ func TestBadCertFile(t *testing.T) {
 	args.CABundleFile = "bad_file.nope"
 
 	_, err := NewClient("")
+	assert.Error(t, err)
+}
+
+func TestAuthRequest(t *testing.T) {
+	// generate a test server so we can capture and inspect the request
+	testServer := httptest.NewServer(http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
+		res.WriteHeader(200)
+		username, password, ok := req.BasicAuth()
+		assert.True(t, ok)
+		assert.Equal(t, username, "testUser")
+		assert.Equal(t, password, "testPass")
+		res.Write([]byte("{\"ok\":true}"))
+	}))
+	defer func() { testServer.Close() }()
+
+	client := &HTTPClient{
+		client:   testServer.Client(),
+		useAuth:  true,
+		username: "testUser",
+		password: "testPass",
+		baseURL:  testServer.URL,
+	}
+
+	testResult := struct {
+		OK *bool `json:"ok"`
+	}{}
+
+	err := client.Request("/endpoint", &testResult)
+	assert.NoError(t, err)
+	assert.Equal(t, true, *testResult.OK)
+}
+
+func TestBadStatusCode(t *testing.T) {
+	testServer := httptest.NewServer(http.HandlerFunc(func(res http.ResponseWriter, req *http.Request) {
+		res.WriteHeader(401)
+		res.Write([]byte("{\"error\":{\"type\":\"exception\",\"reason\":\"this is an error\"}}"))
+	}))
+	defer func() { testServer.Close() }()
+
+	client := &HTTPClient{
+		client:   testServer.Client(),
+		useAuth:  true,
+		username: "testUser",
+		password: "testPass",
+		baseURL:  testServer.URL,
+	}
+
+	err := client.Request("/endpoint", nil)
 	assert.Error(t, err)
 }