+++ src/elasticsearch.go
@@ -10,20 +10,22 @@ import (
 
 type argumentList struct {
 	sdkArgs.DefaultArgumentList
-	Hostname     string `default:"localhost" help:"Hostname or IP where Elasticsearch Node is running."`
-	Port         int    `default:"9200" help:"Port on which Elasticsearch Node is listening."`
-	Username     string `default:"" help:"Username for accessing Elasticsearch Node"`
-	Password     string `default:"" help:"Password for the given user."`
-	UseSSL       bool   `default:"false" help:"Signals whether to use SSL or not. Certificate bundle must be supplied"`
-	CABundleFile string `default:"" help:"Alternative Certificate Authority bundle file"`
-	CABundleDir  string `default:"" help:"Alternative Certificate Authority bundle directory"`
-	Timeout      int    `default:"30" help:"Timeout for an API call"`
-	ConfigPath   string `default:"/etc/elasticsearch/elasticsearch.yml" help:"Path to the ElasticSearch configuration .yml file."`
+	Hostname         string `default:"localhost" help:"Hostname or IP where Elasticsearch Node is running."`
+	Port             int    `default:"9200" help:"Port on which Elasticsearch Node is listening."`
+	Username         string `default:"" help:"Username for accessing Elasticsearch Node"`
+	Password         string `default:"" help:"Password for the given user."`
+	UseSSL           bool   `default:"false" help:"Signals whether to use SSL or not. Certificate bundle must be supplied"`
+	CABundleFile     string `default:"" help:"Alternative Certificate Authority bundle file"`
+	CABundleDir      string `default:"" help:"Alternative Certificate Authority bundle directory"`
+	Timeout          int    `default:"30" help:"Timeout for an API call"`
+	ConfigPath       string `default:"/etc/elasticsearch/elasticsearch.yml" help:"Path to the ElasticSearch configuration .yml file."`
+	CollectIndices   bool   `default:"true" help:"Signals whether to collect indices metrics or not"`
+	CollectPrimaries bool   `default:"true" help:"Signals whether to collect primaries metrics or not"`
 }
 
 const (
 	integrationName    = "com.newrelic.elasticsearch"
-	integrationVersion = "0.1.0"
+	integrationVersion = "0.1.1"
 )
 
 var (