+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.1.1 - 2018-09-13
+### Added
+- Implemented client authentication
+- Implemented toggles for primaries and indices 
+- Added status code and error checking to client requests
+
 ## 0.1.0 - 2018-08-28
 ### Added
 - Initial version: Includes Metrics and Inventory data